refactor(actions): push ref manipulation and dataset version delete into base
