+++ actions/dataset.go
@@ -4,16 +4,13 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"time"
 
 	"github.com/qri-io/dataset"
 	"github.com/qri-io/dataset/dsfs"
-	"github.com/qri-io/dataset/validate"
 	"github.com/qri-io/ioes"
 	"github.com/qri-io/qfs"
 	"github.com/qri-io/qfs/cafs"
 	"github.com/qri-io/qri/base"
-	"github.com/qri-io/qri/logbook"
 	"github.com/qri-io/qri/p2p"
 	"github.com/qri-io/qri/remote"
 	"github.com/qri-io/qri/repo"
@@ -235,77 +232,5 @@ func AddDataset(ctx context.Context, node *p2p.QriNode, rc *remote.Client, remot
 		return fmt.Errorf("error loading added dataset: %s", ref.Path)
 	}
 
-	return ReplaceRefIfMoreRecent(node, &prevRef, ref)
-}
-
-// ReplaceRefIfMoreRecent replaces the given ref in the ref store, if
-// it is more recent then the ref currently in the refstore
-func ReplaceRefIfMoreRecent(node *p2p.QriNode, prev, curr *repo.DatasetRef) error {
-	var (
-		prevTime time.Time
-		currTime time.Time
-	)
-	if curr == nil || curr.Dataset == nil || curr.Dataset.Commit == nil {
-		return fmt.Errorf("added dataset ref is not fully dereferenced")
-	}
-	currTime = curr.Dataset.Commit.Timestamp
-	if prev == nil || prev.Dataset == nil || prev.Dataset.Commit == nil {
-		return fmt.Errorf("previous dataset ref is not fully derefernced")
-	}
-	prevTime = prev.Dataset.Commit.Timestamp
-
-	if prevTime.Before(currTime) || prevTime.Equal(currTime) {
-		if err := node.Repo.PutRef(*curr); err != nil {
-			log.Debug(err.Error())
-			return fmt.Errorf("error putting dataset name in repo: %s", err.Error())
-		}
-	}
-	return nil
-}
-
-// SetPublishStatus configures the publish status of a stored reference
-func SetPublishStatus(node *p2p.QriNode, ref *repo.DatasetRef, published bool) (err error) {
-	if published {
-		node.LocalStreams.PrintErr("📝 listing dataset for p2p discovery\n")
-	} else {
-		node.LocalStreams.PrintErr("unlisting dataset from p2p discovery\n")
-	}
-	return base.SetPublishStatus(node.Repo, ref, published)
-}
-
-// ModifyDataset alters a reference by changing what dataset it refers to
-func ModifyDataset(node *p2p.QriNode, current, new *repo.DatasetRef, isRename bool) (err error) {
-	r := node.Repo
-	if err := validate.ValidName(new.Name); err != nil {
-		return err
-	}
-	if err := repo.CanonicalizeDatasetRef(r, current); err != nil {
-		log.Debug(err.Error())
-		return fmt.Errorf("error with existing reference: %s", err.Error())
-	}
-	err = repo.CanonicalizeDatasetRef(r, new)
-	if err == nil {
-		if isRename {
-			return fmt.Errorf("dataset '%s/%s' already exists", new.Peername, new.Name)
-		}
-	} else if err != repo.ErrNotFound {
-		log.Debug(err.Error())
-		return fmt.Errorf("error with new reference: %s", err.Error())
-	}
-	if isRename {
-		new.Path = current.Path
-
-		if err = r.Logbook().WriteNameAmend(context.TODO(), repo.ConvertToDsref(*current), new.Name); err != nil && err != logbook.ErrNoLogbook {
-			return err
-		}
-	}
-
-	if err = r.DeleteRef(*current); err != nil {
-		return err
-	}
-	if err = r.PutRef(*new); err != nil {
-		return err
-	}
-
-	return nil
+	return base.ReplaceRefIfMoreRecent(node.Repo, &prevRef, ref)
 }
