+++ actions/dataset_test.go
@@ -3,7 +3,6 @@ package actions
 import (
 	"context"
 	"testing"
-	"time"
 
 	"github.com/qri-io/dataset"
 	"github.com/qri-io/qfs"
@@ -27,7 +26,6 @@ func TestAddDataset(t *testing.T) {
 	if err := AddDataset(ctx, node, nil, "", &repo.DatasetRef{Peername: "foo", Name: "bar"}); err == nil {
 		t.Error("expected add of invalid ref to error")
 	}
-
 	// Create test nodes.
 	factory := p2ptest.NewTestNodeFactory(p2p.NewTestableQriNode)
 	testPeers, err := p2ptest.NewTestNetwork(ctx, factory, 2)
@@ -321,13 +319,13 @@ func testRenameDataset(t *testing.T, rmf RepoMakerFunc) {
 		Peername: "me",
 	}
 
-	if err := ModifyDataset(node, &ref, b, true); err != nil {
-		t.Error(err.Error())
+	if err := base.ModifyDatasetRef(ctx, node.Repo, &ref, b, true); err != nil {
+		t.Error(err)
 		return
 	}
 
 	if err := base.ReadDataset(ctx, node.Repo, b); err != nil {
-		t.Error(err.Error())
+		t.Error(err)
 		return
 	}
 
@@ -337,165 +335,12 @@ func testRenameDataset(t *testing.T, rmf RepoMakerFunc) {
 	}
 }
 
-func TestReplaceRefIfMoreRecent(t *testing.T) {
-	node := newTestNode(t)
-	older := time.Date(2019, 1, 1, 12, 0, 0, 0, time.UTC)
-	newer := older.AddDate(1, 0, 0)
-	cases := []struct {
-		description string
-		a, b        repo.DatasetRef
-		path        string
-	}{
-		{
-			"first dataset is older then the second",
-			repo.DatasetRef{
-				Peername:  "woo",
-				Name:      "first_older",
-				Path:      "/map/first",
-				ProfileID: "id",
-				Dataset: &dataset.Dataset{
-					Commit: &dataset.Commit{
-						Timestamp: older,
-					},
-				},
-			},
-			repo.DatasetRef{
-				Peername:  "woo",
-				Name:      "first_older",
-				Path:      "/map/second",
-				ProfileID: "id",
-				Dataset: &dataset.Dataset{
-					Commit: &dataset.Commit{
-						Timestamp: newer,
-					},
-				},
-			},
-			"/map/second",
-		},
-		{
-			"first dataset is newer then the second",
-			repo.DatasetRef{
-				Peername:  "woo",
-				Name:      "first_newer",
-				Path:      "/map/first",
-				ProfileID: "id",
-				Dataset: &dataset.Dataset{
-					Commit: &dataset.Commit{
-						Timestamp: newer,
-					},
-				},
-			},
-			repo.DatasetRef{
-				Peername:  "woo",
-				Name:      "first_newer",
-				Path:      "/map/second",
-				ProfileID: "id",
-				Dataset: &dataset.Dataset{
-					Commit: &dataset.Commit{
-						Timestamp: older,
-					},
-				},
-			},
-			"/map/first",
-		},
-		{
-			"first dataset is same time as the the second",
-			repo.DatasetRef{
-				Peername:  "woo",
-				Name:      "first_same",
-				Path:      "/map/first",
-				ProfileID: "id",
-				Dataset: &dataset.Dataset{
-					Commit: &dataset.Commit{
-						Timestamp: newer,
-					},
-				},
-			},
-			repo.DatasetRef{
-				Peername:  "woo",
-				Name:      "first_same",
-				Path:      "/map/second",
-				ProfileID: "id",
-				Dataset: &dataset.Dataset{
-					Commit: &dataset.Commit{
-						Timestamp: newer,
-					},
-				},
-			},
-			"/map/second",
-		},
-	}
+func testDeleteDataset(t *testing.T, rmf RepoMakerFunc) {
+	ctx := context.Background()
+	node, ref := createDataset(t, rmf)
 
-	for _, c := range cases {
-		if err := node.Repo.PutRef(c.a); err != nil {
-			t.Fatal(err)
-		}
-		if err := ReplaceRefIfMoreRecent(node, &c.a, &c.b); err != nil {
-			t.Fatal(err)
-		}
-		ref, err := node.Repo.GetRef(repo.DatasetRef{Peername: c.a.Peername, Name: c.a.Name})
-		if err != nil {
-			t.Fatal(err)
-		}
-		if ref.Path != c.path {
-			t.Errorf("case '%s', ref path error, expected: '%s', got: '%s'", c.description, c.path, ref.Path)
-		}
-	}
-
-	casesError := []struct {
-		description string
-		a, b        repo.DatasetRef
-		err         string
-	}{
-		{
-			"original ref has no timestamp & should error",
-			repo.DatasetRef{
-				Peername:  "woo",
-				Name:      "err",
-				Path:      "/map/first",
-				ProfileID: "id",
-			},
-			repo.DatasetRef{
-				Peername:  "woo",
-				Name:      "err",
-				Path:      "/map/second",
-				ProfileID: "id",
-				Dataset: &dataset.Dataset{
-					Commit: &dataset.Commit{
-						Timestamp: newer,
-					},
-				},
-			},
-			"previous dataset ref is not fully derefernced",
-		},
-		{
-			"added ref has no timestamp & should error",
-			repo.DatasetRef{
-				Peername:  "woo",
-				Name:      "err",
-				Path:      "/map/first",
-				ProfileID: "id",
-				Dataset: &dataset.Dataset{
-					Commit: &dataset.Commit{
-						Timestamp: newer,
-					},
-				},
-			},
-			repo.DatasetRef{},
-			"added dataset ref is not fully dereferenced",
-		},
-	}
-
-	for _, c := range casesError {
-		if err := node.Repo.PutRef(c.a); err != nil {
-			t.Fatal(err)
-		}
-		err := ReplaceRefIfMoreRecent(node, &c.a, &c.b)
-		if err == nil {
-			t.Errorf("case '%s' did not error", c.description)
-		}
-		if err.Error() != c.err {
-			t.Errorf("case '%s', error mismatch. expected: '%s', got: '%s'", c.description, c.err, err.Error())
-		}
+	if err := base.DeleteDataset(ctx, node.Repo, &ref); err != nil {
+		t.Error(err.Error())
+		return
 	}
 }
