+++ base/ref.go
@@ -1,8 +1,12 @@
 package base
 
 import (
+	"context"
 	"fmt"
+	"time"
 
+	"github.com/qri-io/dataset/validate"
+	"github.com/qri-io/qri/logbook"
 	"github.com/qri-io/qri/repo"
 )
 
@@ -43,4 +47,65 @@ func ToDatasetRef(path string, r repo.Repo, allowFSI bool) (*repo.DatasetRef, er
 		}
 	}
 	return &ref, nil
+}
+
+// ReplaceRefIfMoreRecent replaces the given ref in the ref store, if
+// it is more recent then the ref currently in the refstore
+func ReplaceRefIfMoreRecent(r repo.Repo, prev, curr *repo.DatasetRef) error {
+	var (
+		prevTime time.Time
+		currTime time.Time
+	)
+	if curr == nil || curr.Dataset == nil || curr.Dataset.Commit == nil {
+		return fmt.Errorf("added dataset ref is not fully dereferenced")
+	}
+	currTime = curr.Dataset.Commit.Timestamp
+	if prev == nil || prev.Dataset == nil || prev.Dataset.Commit == nil {
+		return fmt.Errorf("previous dataset ref is not fully derefernced")
+	}
+	prevTime = prev.Dataset.Commit.Timestamp
+
+	if prevTime.Before(currTime) || prevTime.Equal(currTime) {
+		if err := r.PutRef(*curr); err != nil {
+			log.Debug(err.Error())
+			return fmt.Errorf("error putting dataset name in repo: %s", err.Error())
+		}
+	}
+	return nil
+}
+
+// ModifyDatasetRef alters a reference by changing what dataset it refers to
+func ModifyDatasetRef(ctx context.Context, r repo.Repo, current, new *repo.DatasetRef, isRename bool) (err error) {
+	if err := validate.ValidName(new.Name); err != nil {
+		return err
+	}
+	if err := repo.CanonicalizeDatasetRef(r, current); err != nil {
+		log.Debug(err.Error())
+		return fmt.Errorf("error with existing reference: %s", err.Error())
+	}
+	err = repo.CanonicalizeDatasetRef(r, new)
+	if err == nil {
+		if isRename {
+			return fmt.Errorf("dataset '%s/%s' already exists", new.Peername, new.Name)
+		}
+	} else if err != repo.ErrNotFound {
+		log.Debug(err.Error())
+		return fmt.Errorf("error with new reference: %s", err.Error())
+	}
+	if isRename {
+		new.Path = current.Path
+
+		if err = r.Logbook().WriteNameAmend(ctx, repo.ConvertToDsref(*current), new.Name); err != nil && err != logbook.ErrNoLogbook {
+			return err
+		}
+	}
+
+	if err = r.DeleteRef(*current); err != nil {
+		return err
+	}
+	if err = r.PutRef(*new); err != nil {
+		return err
+	}
+
+	return nil
 }
